Added functionality to the GPS class and added a gps testing page to try and visualize if the state being "driven" in is being detected. There is a permission error at the moment. (Can be seen in the 'run' tab when running the app)

+++ app/src/main/java/com/example/electrochromicfilm/Scrollbar.java
@@ -90,8 +90,8 @@ public class Scrollbar extends AppCompatActivity {
         if (btSocket != null) {
             try {
                 // This may not work. May have to turn the tintValue into bytes (next line)
-                //btSocket.getOutputStream().write((byte) tintValue);
-                btSocket.getOutputStream().write(tintValue);
+                btSocket.getOutputStream().write((byte) tintValue);
+                //btSocket.getOutputStream().write(tintValue);
             } catch (IOException e) {
                 Toast.makeText(getApplicationContext(), "Write Attempt Failed", Toast.LENGTH_SHORT).show();
             }
